feature: 무한 스크롤 구현 완료, init: 마이페이지 user type에 따른 컴포넌트를 노출하도록 변경
