+++ src/pages/Home/components/PopularProductList.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { DivBox } from '../../../styles/box';
 import SubTitle from '../../../common/SubTitle';
-import { useProducts } from '../../../api/product/hook/useProducts';
 import { ProductThumbCard } from '../../../common/ProductThumbCard';
+import { useProductsQuery } from '../../../stores/product/useInfiniteProduct';
 function PopularProductList(props) {
-  const { products } = useProducts();
-
+  
+  const { data } = useProductsQuery();
+  const products = data?.pages.flatMap((page) => page.products) || [];
 
   return (
     <>
@@ -15,7 +16,7 @@ function PopularProductList(props) {
           linkText={'더 보기+'}
           link={'/product'}
         />
-        <ProductThumbCard selectdata={products} slice={true} slicecount={4}/>
+        <ProductThumbCard selectdata={products} slice={true} slicecount={4} />
       </DivBox>
     </>
   );
