+++ src/pages/ProductManage/components/CreatorInfoWrap.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import SellProductList from './creator/SellProductList'
+
+export const CreatorInfoWrap = () => {
+  return (
+    <SellProductList />
+  )
+}
