+++ src/pages/ProductManage/components/FanInfoWrap.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const FanInfoWrap = () => {
+  return (
+    <div>FanInfoWrap</div>
+  )
+}
