+++ src/pages/ProductManage/index.jsx
@@ -1,20 +1,30 @@
 import React from 'react';
 import { DivBox } from '../../styles/box';
 import { UserInfoWrap } from './components/UserInfo';
-import SellProductList from './components/SellProductList';
-
+import { CreatorInfoWrap } from './components/CreatorInfoWrap';
+import { FanInfoWrap } from './components/FanInfoWrap';
 import useUserStore from '../../stores/auth/useUserStore';
+
+const COMPONENT_MAP = {
+  creator: <CreatorInfoWrap />,
+  fan: <FanInfoWrap />,
+};
+
 export const ProductManagePage = () => {
- 
+
+  const user = useUserStore((state) => state.user);
+  const Component = COMPONENT_MAP[user.type];
+  // type 이 없는 경우 error 처리
+
   return (
     <>
       <div
         className="product-Manage-page"
         style={{ backgroundColor: '#F5F5F5' }}
       >
-        <UserInfoWrap/>
+        <UserInfoWrap />
         <DivBox>
-          <SellProductList />
+          <Component />
         </DivBox>
       </div>
     </>
